+++ .github/actions/create-project-issue-v1/src/addToBoard.ts
@@ -1,6 +1,5 @@
 import dedent from 'dedent'
 import { getOctokit } from '@actions/github'
-import type { AddProjectCardResponse, ProjectBoardResponse } from './types'
 
 // TODO: Refactor .github/actions/add-to-board to use this: https://github.com/dequelabs/axe-api-team/issues/369
 
@@ -63,6 +62,31 @@ export const MOVE_CARD_TO_COLUMN = dedent`
   }
 `
 
+type Nodes = {
+  id: string
+  name: string
+  options: Omit<Nodes, 'options'>[]
+}
+
+export type ProjectBoardResponse = {
+  organization: {
+    projectV2: {
+      id: string
+      fields: {
+        nodes: Nodes[]
+      }
+    }
+  }
+}
+
+export type AddProjectCardResponse = {
+  addProjectV2ItemById: {
+    item: {
+      id: string
+    }
+  }
+}
+
 interface AddToBoardArgs {
   octokit: ReturnType<typeof getOctokit>
   repositoryOwner: string
@@ -77,37 +101,48 @@ export default async function addToBoard({
   projectNumber,
   columnName,
   issueNodeId
-}: AddToBoardArgs): Promise<void> {
-  const projectBoard = await octokit.graphql<ProjectBoardResponse>(
-    GET_PROJECT_BOARD_BY_NUMBER,
-    {
-      owner: repositoryOwner,
-      projectNumber
-    }
-  )
+}: AddToBoardArgs): Promise<any> {
+  try {
+    const projectBoard = await octokit.graphql<ProjectBoardResponse>(
+      GET_PROJECT_BOARD_BY_NUMBER,
+      {
+        owner: repositoryOwner,
+        projectNumber
+      }
+    )
 
-  const projectCard = await octokit.graphql<AddProjectCardResponse>(
-    ADD_ISSUE_TO_PROJECT_BOARD,
-    {
+    const projectCard = await octokit.graphql<AddProjectCardResponse>(
+      ADD_ISSUE_TO_PROJECT_BOARD,
+      {
+        projectId: projectBoard.organization.projectV2.id,
+        issueId: issueNodeId
+      }
+    )
+
+    // Status is the array of columns like Backlog, In Progress, Done etc
+    //eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+    const statusColumn = projectBoard.organization.projectV2.fields.nodes.find(
+      node => node.name === 'Status'
+    )!
+
+    //eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+    const column = statusColumn.options.find(
+      option => option.name.toLowerCase() === columnName.toLowerCase()
+    )
+
+    if (!column) {
+      throw new Error(`Column ${columnName} not found`)
+    }
+
+    const response = await octokit.graphql(MOVE_CARD_TO_COLUMN, {
       projectId: projectBoard.organization.projectV2.id,
-      issueId: issueNodeId
-    }
-  )
+      itemId: projectCard.addProjectV2ItemById.item.id,
+      fieldId: statusColumn.id,
+      value: column.id
+    })
 
-  //eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-  const statusColumn = projectBoard.organization.projectV2.fields.nodes.find(
-    node => node.name === 'Status'
-  )!
-
-  //eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-  const column = statusColumn.options.find(
-    option => option.name.toLowerCase() === columnName.toLowerCase()
-  )!
-
-  await octokit.graphql(MOVE_CARD_TO_COLUMN, {
-    projectId: projectBoard.organization.projectV2.id,
-    itemId: projectCard.addProjectV2ItemById.item.id,
-    fieldId: statusColumn.id,
-    value: column.id
-  })
+    return response
+  } catch (error) {
+    throw new Error(`Add to board failed: ${(error as Error).message}`)
+  }
 }
