+++ .github/actions/create-project-issue-v1/src/types.ts
@@ -6,28 +6,3 @@ export type Core = Pick<
   'getInput' | 'setFailed' | 'info' | 'warning' | 'setOutput'
 >
 export type GitHub = Pick<typeof github, 'getOctokit' | 'context'>
-
-type Nodes = {
-  id: string
-  name: string
-  options: Omit<Nodes, 'options'>[]
-}
-
-export type ProjectBoardResponse = {
-  organization: {
-    projectV2: {
-      id: string
-      fields: {
-        nodes: Nodes[]
-      }
-    }
-  }
-}
-
-export type AddProjectCardResponse = {
-  addProjectV2ItemById: {
-    item: {
-      id: string
-    }
-  }
-}
